Add pull command to CLI
Fix admin html ng-model
