+++ cmd/evergreen/evergreen.go
@@ -53,6 +53,7 @@ func buildApp() *cli.App {
 		// Top-level commands.
 		operations.Keys(),
 		operations.Fetch(),
+		operations.Pull(),
 		operations.Evaluate(),
 		operations.Validate(),
 		operations.List(),
