+++ operations/fetch.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"runtime"
 	"strconv"
 	"strings"
 	"sync"
@@ -23,7 +22,6 @@ import (
 	"github.com/evergreen-ci/evergreen/service"
 	"github.com/evergreen-ci/evergreen/thirdparty"
 	"github.com/evergreen-ci/evergreen/util"
-	"github.com/evergreen-ci/pail"
 	"github.com/mongodb/grip"
 	"github.com/mongodb/grip/message"
 	"github.com/pkg/errors"
@@ -47,9 +45,6 @@ func Fetch() cli.Command {
 	return cli.Command{
 		Name:  "fetch",
 		Usage: "fetch the source or artifacts associated with a task",
-		Subcommands: []cli.Command{
-			fetchFromPull(),
-		},
 		Flags: []cli.Flag{
 			cli.StringFlag{
 				Name:  joinFlagNames(dirFlagName, "d"),
@@ -598,78 +593,4 @@ func downloadUrls(root string, urls chan artifactDownload, workers int) error {
 	<-done
 
 	return hasErrors
-}
-
-func fetchFromPull() cli.Command {
-	const (
-		taskFlagName = "task"
-		dirFlagName  = "dir"
-	)
-	return cli.Command{
-		Name:  "pull",
-		Usage: "pull a completed task's exact working directory contents",
-		Flags: []cli.Flag{
-			cli.StringFlag{
-				Name:  joinFlagNames(taskFlagName, "t"),
-				Usage: "the ID of the task to pull",
-			},
-			cli.StringFlag{
-				Name:  joinFlagNames(dirFlagName, "d"),
-				Usage: "the directory to put the task contents (default: current working directory)",
-			},
-		},
-		Before: mergeBeforeFuncs(
-			requireStringFlag(taskFlagName),
-			requireWorkingDirFlag(dirFlagName),
-		),
-		Action: func(c *cli.Context) error {
-			confPath := c.Parent().Parent().String(confFlagName)
-			taskID := c.String(taskFlagName)
-			workingDir := c.String(dirFlagName)
-
-			conf, err := NewClientSettings(confPath)
-			if err != nil {
-				return errors.Wrap(err, "problem loading configuration")
-			}
-
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			client := conf.setupRestCommunicator(ctx)
-			defer client.Close()
-
-			creds, err := client.GetS3TaskCredentials(ctx)
-			if err != nil {
-				return errors.Wrap(err, "could not fetch credentials")
-			}
-
-			s3Path, err := client.GetS3TaskPath(ctx, taskID)
-			if err != nil {
-				return errors.Wrap(err, "could not get location of task directory in S3")
-			}
-
-			httpClient := util.GetHTTPClient()
-			defer util.PutHTTPClient(httpClient)
-			bucket, err := pail.NewS3MultiPartBucketWithHTTPClient(httpClient, pail.S3Options{
-				Credentials: pail.CreateAWSCredentials(creds.Key, creds.Secret, ""),
-				// kim: TODO: validate correctness of permissions once
-				// credentials are received.
-				Permissions: pail.S3PermissionsAWSExecRead,
-				Verbose:     true,
-			})
-			if err != nil {
-				return errors.Wrap(err, "error setting up S3 bucket")
-			}
-			bucket = pail.NewParallelSyncBucket(pail.ParallelBucketOptions{
-				Workers: runtime.NumCPU(),
-			}, bucket)
-
-			bucket.Pull(ctx, pail.SyncOptions{
-				Local:  workingDir,
-				Remote: s3Path,
-			})
-
-			return nil
-		},
-	}
 }
