+++ service/templates/admin.html
@@ -1392,39 +1392,39 @@ Admin Settings
 
 									<md-input-container class="control" style="width:45%;">
 										<label>S3 Key</label>
-										<input type="text" ng-model="Settings.providers.aws.s3.key">
+										<input type="text" ng-model="Settings.providers.aws.s3_credentials.key">
 									</md-input-container>
 									<md-input-container class="control" style="width:45%; margin-left:50px;">
 										<label>S3 Secret</label>
-										<input type="text" ng-model="Settings.providers.aws.s3.secret">
+										<input type="text" ng-model="Settings.providers.aws.s3_credentials.secret">
 									</md-input-container>
 									<md-input-container class="control" style="width:45%;">
 										<label>S3 Bucket</label>
-										<input type="text" ng-model="Settings.providers.aws.s3.bucket">
+										<input type="text" ng-model="Settings.providers.aws.s3_credentials.bucket">
 									</md-input-container>
 									<md-input-container class="control" style="width:45%; margin-left:50px;">
 										<label>Task S3 Key</label>
-										<input type="text" ng-model="Settings.providers.aws.s3_task.key">
+										<input type="text" ng-model="Settings.providers.aws.s3_task_credentials.key">
 									</md-input-container>
 									<md-input-container class="control" style="width:45%;">
 										<label>Task S3 Secret</label>
-										<input type="text" ng-model="Settings.providers.aws.s3_task.secret">
+										<input type="text" ng-model="Settings.providers.aws.s3_task_credentials.secret">
 									</md-input-container>
 									<md-input-container class="control" style="width:45%; margin-left:50px;">
 										<label>Task S3 Bucket</label>
-										<input type="text" ng-model="Settings.providers.aws.s3_task.bucket">
+										<input type="text" ng-model="Settings.providers.aws.s3_task_credentials.bucket">
 									</md-input-container>
 									<md-input-container class="control" style="width:45%;">
 										<label>Task Read-Only S3 Key</label>
-										<input type="text" ng-model="Settings.providers.aws.s3_task_read.key">
+										<input type="text" ng-model="Settings.providers.aws.s3_task_read_credentials.key">
 									</md-input-container>
 									<md-input-container class="control" style="width:45%; margin-left:50px;">
 										<label>Task Read-Only S3 Secret</label>
-										<input type="text" ng-model="Settings.providers.aws.s3_task_read.secret">
+										<input type="text" ng-model="Settings.providers.aws.s3_task_read_credentials.secret">
 									</md-input-container>
 									<md-input-container class="control" style="width:45%;">
 										<label>Task Read-Only S3 Bucket</label>
-										<input type="text" ng-model="Settings.providers.aws.s3_task_read.bucket">
+										<input type="text" ng-model="Settings.providers.aws.s3_task_read_credentials.bucket">
 									</md-input-container>
 									<md-input-container class="control" style="width:45%; margin-left:50px;">
 										<label>S3 Base URL</label>
